code: move worker into this project again
Having a seperate worker on heroku is a bit of a PITA. A seperate dyno
is required that would have to be pinged to keep it awake. So the plan
now is to add the worker in this project while on heroku, and pull in
the worker code separately via npm.