code: fix permissions issue
Previously, a user could create a link between an upstream they didn't
own and all of its forks, while having no relationship with the
upstream repository or maintainer. This is bad. Also, the same thing
could happen with someone creating a link between an upstream and a fork
they didn't own.

When adding a link, check to make sure the user has permission:

- If the fork.type == 'fork-all', then make sure the user is a
collaborator on the upstream. The thinking here is that only people that
can push code to the upstream should have the power to update how
Backstroke works when changes are synced to all forks.

- If the fork.type == 'repo', then make sure the user is a collaborator
on the fork that changes will be synced into. This is so that
unauthorized users can't sync changes into a fork that you own.