code: fix serious auth issue
I made the assumption that calling `User.update` returns the id of the
updated item. In fact, it returns the number of items successfully
updated. Because of this, repeat logins to backstroke would cause users
to be logged into the accunt of user id `1`, which just so happened to
be me!

Thanks to @brysgo for reporting. 1egoman/backstroke#67